reordering and stabilization
