+++ sim/blop_sim/backends/__init__.py
@@ -2,6 +2,7 @@
 
 from .core import SimBackend
 from .simple import SimpleBackend
-from .xrt import XRTBackend
+from .xrt import XRTBackend, build_histRGB
+from .models.xrt_kb_model import build_beamline, run_process
 
-__all__ = ["SimBackend", "SimpleBackend", "XRTBackend"]
+__all__ = ["SimBackend", "SimpleBackend", "XRTBackend", "build_histRGB", "build_beamline", "run_process"]
