+++ sim/blop_sim/backends/core.py
@@ -14,8 +14,11 @@ class SimBackend(ABC):
 
     _instances: dict[type, "SimBackend"] = {}
 
-    def __new__(cls):
-        """Singleton pattern: return existing instance or create new."""
+    def __new__(cls, *args, **kwargs):
+        """Singleton pattern: return existing instance or create new.
+        Accept arbitrary constructor args/kwargs so subclass constructors
+        (for example XRTBackend(fileName=...)) do not fail at __new__.
+        """
         if cls not in cls._instances:
             instance = super().__new__(cls)
             cls._instances[cls] = instance
