+++ sim/blop_sim/backends/xrt.py
@@ -8,77 +8,35 @@ import xrt.backends.raycing as raycing
 from bluesky.protocols import Readable
 from xrt.backends.raycing import BeamLine
 
-from ..devices.xrt import element_to_variables, i
-from . import SimBackend
-from .models.xrt_kb_model import build_beamline, build_histRGB, run_process
+from .core import SimBackend
+from blop.protocols import MovableHasName, Readable
+from pathlib import Path
 
 
-class KBBackend(SimBackend):
-    """XRT ray-tracing simulation backend.
+limits = [[-0.6, 0.6], [-0.45, 0.45]]
 
-    Uses the XRT package to perform realistic ray-tracing through a KB mirror pair.
-    Much slower than SimpleBackend but more physically accurate.
-    """
 
-    def __init__(self, noise: bool = False):
-        """Initialize XRT backend."""
-        super().__init__()
-        self._beamline = None
-        self._limits = [[-0.6, 0.6], [-0.45, 0.45]]
-        self._noise = noise
+def build_histRGB(lb, gb, limits=None, isScreen=False, shape=None):
+    if shape is None:
+        shape = [256, 256]
+    good = (lb.state == 1) | (lb.state == 2)
+    if isScreen:
+        x, y, z = lb.x[good], lb.z[good], lb.y[good]
+    else:
+        x, y, z = lb.x[good], lb.y[good], lb.z[good]
+    goodlen = len(lb.x[good])
+    hist2dRGB = np.zeros((shape[1], shape[0], 3), dtype=np.float64)
+    hist2d = np.zeros((shape[1], shape[0]), dtype=np.float64)
 
-    def _ensure_beamline(self):
-        """Build XRT beamline if not already built."""
-        if self._beamline is None:
-            self._beamline = build_beamline()
+    if limits is None and goodlen > 0:
+        limits = np.array([[np.min(x), np.max(x)], [np.min(y), np.max(y)], [np.min(z), np.max(z)]])
 
-    async def generate_beam(self) -> np.ndarray:
-        """Generate beam using XRT ray-tracing.
-
-        Returns:
-            2D numpy array with shape (300, 400)
-        """
-        self._ensure_beamline()
-
-        # Get KB mirror radii from devices
-        mirror_radii = await self._get_mirror_radii()
-
-        # Update XRT beamline mirror parameters
-        self._beamline.toroidMirror01.R = mirror_radii[0]  # Vertical mirror
-        self._beamline.toroidMirror02.R = mirror_radii[1]  # Horizontal mirror
-
-        # Run ray tracing
-        outDict = run_process(self._beamline)
-        lb = outDict["screen01beamLocal01"]
-
-        # Build histogram from ray data
-        hist2d, _, _ = build_histRGB(lb, lb, limits=self._limits, isScreen=True, shape=[400, 300])
-        image = hist2d
-
-        # Add noise if requested
-        if self._noise:
-            image += 1e-3 * np.abs(np.random.standard_normal(size=image.shape))
-
-        return image
-
-    async def _get_mirror_radii(self) -> list[float]:
-        """Get KB mirror radii from registered devices.
-
-        Returns:
-            [R1, R2] where R1 is first mirror (vertical), R2 is second mirror (horizontal)
-        """
-        # Default radii from xrt_kb_model.py
-        radii = [38245.0, 21035.0]
-
-        for name, device in self._device_states.items():
-            if device["type"] == "kb_mirror_xrt":
-                state = await self._get_device_state(name)
-                mirror_index = state["mirror_index"]
-                radius = state["radius"]
-                if mirror_index < len(radii):
-                    radii[mirror_index] = radius
-
-        return radii
+    if goodlen > 0:
+        beamLimits = [limits[1], limits[0]] or None
+        flux = gb.Jss[good] + gb.Jpp[good]
+        hist2d, _, _ = np.histogram2d(y, x, bins=[shape[1], shape[0]], range=beamLimits, weights=flux)
+        hist2dRGB = None
+    return hist2d, hist2dRGB, limits
 
 
 class XRTBackend(SimBackend, Readable):
@@ -88,13 +46,16 @@ class XRTBackend(SimBackend, Readable):
     Much slower than SimpleBackend but more physically accurate.
     """
 
-    def __init__(self, file=None, limits=None, noise: bool = False):
+    def __init__(self, file, limits=None, noise: bool = False):
         """Initialize XRT backend."""
         super().__init__()
         self._beamline = raycing.BeamLine(fileName=file)
+        self._name = Path(file).stem
         self._limits = limits or [[-0.6, 0.6], [-0.45, 0.45]]
         self._noise = noise
         self._target = None
+        beamLine = self._beamline
+        self._elements = {k: beamLine.oesDict[v][0] for k, v in beamLine.oenamesToUUIDs.items()}
         self._ensure_beamline()
 
     def _ensure_beamline(self):
@@ -102,10 +63,6 @@ class XRTBackend(SimBackend, Readable):
         if self._beamline is None:
             raise ValueError("beamline has not been initialized")
 
-        if self._elements is None:
-            beamLine = self._beamline
-            self._elements = {k: beamLine.oesDict[v][0] for k, v in beamLine.oenamesToUUIDs.items()}
-            self._variables = {k: element_to_variables(v, k) for k, v in self._elements.items()}
 
     def generate_beam(self) -> np.ndarray:
         """Generate beam using XRT ray-tracing.
@@ -119,9 +76,9 @@ class XRTBackend(SimBackend, Readable):
         outDict = raycing.run_process_from_file(self._beamline)
         self.render = outDict
         if self.target is not None:
-            print("warning: target is not set, please make sure you manage your triggering by setting a primary detector")
-            lb = [v for k, v in outDict.items() if self.target in k][0]
+            lb = [v for k, v in outDict.items() if self.target.name in k][0]
         else:
+            print("warning: target is not set, please make sure you manage your triggering by setting a primary detector")
             lb = outDict.values()[0]
 
         # Build histogram from ray data
@@ -139,7 +96,7 @@ class XRTBackend(SimBackend, Readable):
         return self._target
 
     @target.setter
-    def register_target(self, detector):
+    def target(self, detector):
         self._target = detector
 
     def read(self):
@@ -154,12 +111,16 @@ class XRTBackend(SimBackend, Readable):
     def describe(self):
         return OrderedDict([(k, {"source": k, "dtype": type(v['value']), 'shape': v.shape}) for k, v in self.read().items()])
 
+    @property
+    def name(self):
+        return self._name
+
     def __getitem__(self, key):
         if self.render is None:
             self.generate_beam()
         if key in self.render.keys():
-            return self.render[key]
-        return [v for k, v in self.render.items() if key in k and 'local' in k]
+            return [self.render[key]]
+        return [k: v for k, v in self.render.items() if key in k and 'local' in k]
 
     @property
     def variables(self):
@@ -168,3 +129,4 @@ class XRTBackend(SimBackend, Readable):
     @property
     def elements(self):
         return self._elements
+
